requirements.txt
ultralytics==8.3.167
opencv-python==4.12.0.88
gradio==5.42.0
pillow==11.3.0
rembg[cpu]==2.0.67
scikit-learn==1.7.2
pandas==2.3.2
pyyaml==6.0.2